debugging.soft: Support breakpoint and stacktrace in iterator methods
When a method is compiled as an iterator (i.e. `yield return`) it will
result in the creation of a compiler-generated type and will not have
any location information of its own.  This leads to breakpoints and
stack traces to not work as the adapter cannot map file locations to the
appropraite IL offsets.

This patch tries to detect those cases, and redirects to the search into
the appropriate generated types.  Breaking inside an interator method
now work.

However, it is currently not possible to step through a `yield`
statement to have the runtime stop on the next line.